scraper/requirements.txt
playwright>=1.40
python-dotenv>=1.0
pytest>=7.0